RF+BF: Move reporting code to own subpackage, break up big modules
Few fixes/improvements around reporting. In particular stop raising
exceptions in the setter for the done propety.
